requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-playwright==0.4.1
pytest-asyncio==0.23.8
playwright==1.40.0

# Configuration Management
python-dotenv==1.0.0

# Reporting
allure-pytest==2.13.2

# Optional: Parallel execution and retry
pytest-xdist==3.5.0
pytest-rerunfailures==12.0

# Optional: Code quality
black==23.12.0
flake8==6.1.0
